0002: rename solutions and strip down files
